[enhance] Speed up package installation
Turns out, the documention for Ansible was wrong, and that optimization
of a with_items list sent to the package (or yum) module does not happen
as documented.

In order to not have installation of packages done one at a time (which
instantiates an instance of yum for every row, making installation slow)
you need to pass the list of items to Ansible as a comma separated list.

In the future, the intention is to be able to provide a list directly
to the name argument, but that isn't implemented yet.